Changes for review comments [HZ-1459]
